fix: Tags in recurrent event not created
