+++ lib/core/database/services/tags/tags_service.dart
@@ -20,6 +20,23 @@ class TagService {
     return (db.delete(db.tags)..where((tbl) => tbl.id.equals(tagId))).go();
   }
 
+  Future<void> linkTagsToTransaction({
+    required String transactionId,
+    required List<String> tagIds,
+  }) {
+    final db = AppDB.instance;
+
+    return db.batch((batch) {
+      for (final tagId in tagIds) {
+        batch.insert(
+          db.transactionTags,
+          TransactionTag(transactionID: transactionId, tagID: tagId),
+          mode: InsertMode.insertOrIgnore,
+        );
+      }
+    });
+  }
+
   Stream<List<Tag>> getTags({
     Expression<bool> Function(Tags)? filter,
     int? limit,
